Fix proposition for the potential error on the creation of a powerplatform_billing_policy resource when receiving a non-final status (#299)
* feat: Add support for waiting for billing policy to reach a terminal state

The code changes introduce a new function `DoWaitForLifecycleOperationStatus` in the `LicensingClient` struct. This function allows the client to wait for a billing policy to reach a terminal state (either "Enabled" or "Disabled") before returning the policy. The function uses a loop with a timeout to periodically check the status of the billing policy and waits for it to reach a terminal state.

* Change the name of the new method to "DoWaitForFinalStatus"

* Remove comments based on feedback in PR

* Add //lintignore:R018 above time.Sleep(retryAfter) as recommended

* Use fmt.Sprintf(..) instead of string concatenation

* Reduction of the duration for the timeout for the creation of a billing policy because 10 minutes it was too long for the unit tests

* Add new unit tests for billing policies

---------

Co-authored-by: mawasile <50197777+mawasile@users.noreply.github.com>